requirements.txt
# USPTO Multi-Processor Requirements
# Install with: pip install -r requirements.txt

# Core dependencies
requests>=2.28.0
pandas>=1.5.0
psycopg2-binary>=2.9.0
lxml>=4.9.0

# Optional dependencies for better performance
numpy>=1.21.0
python-dateutil>=2.8.0

# Development dependencies (optional)
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0